working >rpg info

+++ PythonBot/rpggame/rpgshopitem.py
@@ -1,7 +1,8 @@
 class RPGShopItem:
-    def __init__(self, name : str, cost : int, benefit : {str : (str, float)}):
+    def __init__(self, name: str, cost: float, benefit: {str: (str, float)}):
         self.name = name
-        self.cost = cost # Minutes or Money
+        self.cost = cost  # Minutes or Money
         self.benefit = benefit
+
     def __str__(self):
         return "{}, ${} : {}".format(self.name, self.cost, self.benefit)
